// packages/astro-loader-github-releases/src/graphql/gen/operations.ts
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A (potentially binary) string encoded using base64. */
  Base64String: { input: any; output: any }
  /**
   * Represents non-fractional signed whole numeric values. Since the value may
   * exceed the size of a 32-bit integer, it's encoded as a string.
   */
  BigInt: { input: any; output: any }
  /** An ISO-8601 encoded date string. */
  Date: { input: any; output: any }
  /** An ISO-8601 encoded UTC date string. */
  DateTime: { input: string; output: string }
  /** A Git object ID. */
  GitObjectID: { input: any; output: any }
  /** A fully qualified reference name (e.g. `refs/heads/master`). */
  GitRefname: { input: any; output: any }
  /** Git SSH string */
  GitSSHRemote: { input: any; output: any }
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: { input: any; output: any }
  /** A string containing HTML code. */
  HTML: { input: string; output: string }
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: { input: any; output: any }
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: { input: string; output: string }
  /** A valid x509 certificate string */
  X509Certificate: { input: any; output: any }
}
// ...
export type GetReleaseByTagNameQueryVariables = Exact<{
  owner: Scalars['String']['input']
  repo: Scalars['String']['input']
  tagName: Scalars['String']['input']
}>

export type GetReleaseByTagNameQuery = {
  repository?: {
    release?: {
      id: string
      url: string
      name?: string | null
      tagName: string
      description?: string | null
      descriptionHTML?: string | null
      isDraft: boolean
      isLatest: boolean
      isPrerelease: boolean
      createdAt: string
      publishedAt?: string | null
      repository: {
        name: string
        nameWithOwner: string
        url: string
        stargazerCount: number
        isInOrganization: boolean
      }
    } | null
  } | null
}

export type GetReleaseByIdQueryVariables = Exact<{
  id: Scalars['ID']['input']
}>

export type GetReleaseByIdQuery = {
  node?:
    | {
        id: string
        url: string
        name?: string | null
        tagName: string
        description?: string | null
        descriptionHTML?: string | null
        isDraft: boolean
        isLatest: boolean
        isPrerelease: boolean
        createdAt: string
        publishedAt?: string | null
        repository: {
          name: string
          nameWithOwner: string
          url: string
          stargazerCount: number
          isInOrganization: boolean
        }
      }
    | {}
    | null
}
// ...
export const GetReleaseByTagNameDocument = {
  kind: 'Document',
  definitions: [
    {
      kind: 'OperationDefinition',
      operation: 'query',
      name: { kind: 'Name', value: 'GetReleaseByTagName' },
      variableDefinitions: [
        {
          kind: 'VariableDefinition',
          variable: {
            kind: 'Variable',
            name: { kind: 'Name', value: 'owner' },
          },
          type: {
            kind: 'NonNullType',
            type: {
              kind: 'NamedType',
              name: { kind: 'Name', value: 'String' },
            },
          },
        },
        {
          kind: 'VariableDefinition',
          variable: { kind: 'Variable', name: { kind: 'Name', value: 'repo' } },
          type: {
            kind: 'NonNullType',
            type: {
              kind: 'NamedType',
              name: { kind: 'Name', value: 'String' },
            },
          },
        },
        {
          kind: 'VariableDefinition',
          variable: {
            kind: 'Variable',
            name: { kind: 'Name', value: 'tagName' },
          },
          type: {
            kind: 'NonNullType',
            type: {
              kind: 'NamedType',
              name: { kind: 'Name', value: 'String' },
            },
          },
        },
      ],
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'repository' },
            arguments: [
              {
                kind: 'Argument',
                name: { kind: 'Name', value: 'owner' },
                value: {
                  kind: 'Variable',
                  name: { kind: 'Name', value: 'owner' },
                },
              },
              {
                kind: 'Argument',
                name: { kind: 'Name', value: 'name' },
                value: {
                  kind: 'Variable',
                  name: { kind: 'Name', value: 'repo' },
                },
              },
            ],
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'release' },
                  arguments: [
                    {
                      kind: 'Argument',
                      name: { kind: 'Name', value: 'tagName' },
                      value: {
                        kind: 'Variable',
                        name: { kind: 'Name', value: 'tagName' },
                      },
                    },
                  ],
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      {
                        kind: 'FragmentSpread',
                        name: { kind: 'Name', value: 'ReleaseCore' },
                      },
                    ],
                  },
                },
              ],
            },
          },
        ],
      },
    },
    {
      kind: 'FragmentDefinition',
      name: { kind: 'Name', value: 'ReleaseCore' },
      typeCondition: {
        kind: 'NamedType',
        name: { kind: 'Name', value: 'Release' },
      },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          { kind: 'Field', name: { kind: 'Name', value: 'id' } },
          { kind: 'Field', name: { kind: 'Name', value: 'url' } },
          { kind: 'Field', name: { kind: 'Name', value: 'name' } },
          { kind: 'Field', name: { kind: 'Name', value: 'tagName' } },
          { kind: 'Field', name: { kind: 'Name', value: 'description' } },
          { kind: 'Field', name: { kind: 'Name', value: 'descriptionHTML' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isDraft' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isLatest' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isPrerelease' } },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'repository' },
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                { kind: 'Field', name: { kind: 'Name', value: 'name' } },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'nameWithOwner' },
                },
                { kind: 'Field', name: { kind: 'Name', value: 'url' } },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'stargazerCount' },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'isInOrganization' },
                },
              ],
            },
          },
          { kind: 'Field', name: { kind: 'Name', value: 'createdAt' } },
          { kind: 'Field', name: { kind: 'Name', value: 'publishedAt' } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<
  GetReleaseByTagNameQuery,
  GetReleaseByTagNameQueryVariables
>
export const GetReleaseByIdDocument = {
  kind: 'Document',
  definitions: [
    {
      kind: 'OperationDefinition',
      operation: 'query',
      name: { kind: 'Name', value: 'GetReleaseById' },
      variableDefinitions: [
        {
          kind: 'VariableDefinition',
          variable: { kind: 'Variable', name: { kind: 'Name', value: 'id' } },
          type: {
            kind: 'NonNullType',
            type: { kind: 'NamedType', name: { kind: 'Name', value: 'ID' } },
          },
        },
      ],
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'node' },
            arguments: [
              {
                kind: 'Argument',
                name: { kind: 'Name', value: 'id' },
                value: {
                  kind: 'Variable',
                  name: { kind: 'Name', value: 'id' },
                },
              },
            ],
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'FragmentSpread',
                  name: { kind: 'Name', value: 'ReleaseCore' },
                },
              ],
            },
          },
        ],
      },
    },
    {
      kind: 'FragmentDefinition',
      name: { kind: 'Name', value: 'ReleaseCore' },
      typeCondition: {
        kind: 'NamedType',
        name: { kind: 'Name', value: 'Release' },
      },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          { kind: 'Field', name: { kind: 'Name', value: 'id' } },
          { kind: 'Field', name: { kind: 'Name', value: 'url' } },
          { kind: 'Field', name: { kind: 'Name', value: 'name' } },
          { kind: 'Field', name: { kind: 'Name', value: 'tagName' } },
          { kind: 'Field', name: { kind: 'Name', value: 'description' } },
          { kind: 'Field', name: { kind: 'Name', value: 'descriptionHTML' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isDraft' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isLatest' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isPrerelease' } },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'repository' },
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                { kind: 'Field', name: { kind: 'Name', value: 'name' } },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'nameWithOwner' },
                },
                { kind: 'Field', name: { kind: 'Name', value: 'url' } },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'stargazerCount' },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'isInOrganization' },
                },
              ],
            },
          },
          { kind: 'Field', name: { kind: 'Name', value: 'createdAt' } },
          { kind: 'Field', name: { kind: 'Name', value: 'publishedAt' } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<GetReleaseByIdQuery, GetReleaseByIdQueryVariables>